Mudança do tipo de dado prioridade.. de Int para Long

+++ heap/estatico/Kotlin/Arvore Binária Heap Máximo Paciente/Paciente.kt
@@ -1,1 +1,1 @@
-data class Paciente(val nome: String, val idade: Int, val prioridade: Int)
+data class Paciente(val nome: String, val idade: Int, val prioridade: Long)
